+++ manualCreditor.go
@@ -42,8 +42,13 @@ func (mc *manualCreditor) FlowBits(currentCredits uint32) (bool, uint32) {
 	mc.mu.Lock()
 	defer mc.mu.Unlock()
 
-	drain := mc.drained != nil
+	drain := mc.pendingDrain
 	var credits uint32
+
+	if mc.pendingDrain {
+		// only send one drain request
+		mc.pendingDrain = false
+	}
 
 	// either:
 	// drain is true (ie, we're going to send a drain frame, and the credits for it should be 0)
@@ -55,7 +60,6 @@ func (mc *manualCreditor) FlowBits(currentCredits uint32) (bool, uint32) {
 	}
 
 	mc.creditsToAdd = 0
-	mc.pendingDrain = false
 
 	return drain, credits
 }
@@ -72,6 +76,7 @@ func (mc *manualCreditor) Drain(ctx context.Context, l *link) error {
 	mc.drained = make(chan struct{})
 	// use a local copy to avoid racing with EndDrain()
 	drained := mc.drained
+	mc.pendingDrain = true
 
 	mc.mu.Unlock()